refactor: Reestruturando o axios para consumo da api
