+++ src/services/index.js
@@ -1,0 +1,3 @@
+export * from "./api";
+export * from "./students";
+
