+++ src/services/students/index.js
@@ -1,0 +1,1 @@
+export * from "./students.service";
